Use a memcmp for the expected symbol name. (#100)
* Use a memcmp for the expected symbol name.

I believe that g++ does not guarantee what a particular
symbol name will be.  Thus, in g++ 11.4.0 (what is in
Ubuntu 22.04), the symbol name here ended with "#2", while
in g++ 13.2.0 (what is in Ubuntu 24.04), the symbol name
ends with "#1".  Given that we can't guarantee this, just
search for the first part of the name up to the number,
which should be good enough for this test.

Signed-off-by: Chris Lalancette <clalancette@gmail.com>