Fix up two different C problems.
The first one was that the visibility_control file was
marked as a C++ header, but was included in a C header.
Avoid that by just making visibility_control a C header.

The second problem is that, in C, a function with no
arguments technically takes "any" arguments.  Fix
ros_trace_compile_status() to have void arguments, which
ensures that no arguments can be passed.

Signed-off-by: Chris Lalancette <clalancette@gmail.com>