Change expected rmw GID array size to 16 bytes
The size changed from 24 to 16 bytes in ros2/rmw#345. The same size is
defined in `tracetools`, and it was never updated.

In practice, reading 24-16=8 random extra bytes didn't change much,
since nothing is currently relying on the GID values, at least not
relying on getting the same GID for the same object from two different
systems, because it doesn't work with `rmw_cyclonedds`, see
ros2/rmw_cyclonedds#377.

Signed-off-by: Christophe Bedard <christophe.bedard@apex.ai>